Aggregation-class	R Documentation

A aggregation method reduces the performance values of the test (and possibly the training sets) to a single...

Description

A aggregation method reduces the performance values of the test (and possibly the training sets) to a single value. To see all possible, implemented aggregations look at aggregations.

Details

The aggregation can access all relevant information of the result after resampling and combine them into a single value. Though usually something very simple like taking the mean of the test set performances is done.

Slots

id:

(character) Name of aggregation function.

fun:

(function) Aggregation function: function(perf.test, perf.train, measure, group, pred)), where data types of arguments are: numeric, numeric, Measure, factor, ResamplePrediction.

Methods

setAggregation

See Also

aggregations, Measure, measures