BaseWrapper-class	R Documentation

Abstract base class to wrap something around a learner.

Description

Abstract base class to wrap something around a learner.

Extends

Learner

Methods

trainLearner

predictLearner

getParameterSet

setHyperPars

getHyperPars