LearnTask-class	R Documentation

Base class for learning tasks.

Description

A machine learning task. It encapsulates the data and specifies - through its subclasses - the type of the task (either classification or regression), the target variable and other details of the problem.

Slots

dataenv:

(environment) Environment where data for the task are stored. Use getData in order to access the data.

weights:

(numeric) Case weights. numeric(0) if no weights were set.

blocking:

(factor) Observations with the same blocking level "belong together". Specifically, they are either put all in the training or the test set during a resampling iteration. factor(0) if no blocking was set.

desc:

(TaskDesc) An object of class TaskDesc which encapsulates the main information about the task.

Subclasses: ClassifTask, RegrTask

Methods

trainLearner

subsetData

getFeatureNames

getTargets

getFormula

makeResampleInstance

getLearners

performance

resample

train

See Also

makeClassifTask, makeRegrTask