Learner-class	R Documentation

Base class for inducers.

Description

Abstract base class for learning algorithms.

Details

How to change object later on: Look at setters.

Tresholds for class labels: If you set predict.type to "prob", the label with the maximum value is selected. You can change labels of a Prediction object later by using the function setThreshold.

Getter.

is.classif [logical(1)]

Is this learner for classification tasks?

is.regr [logical(1)]

Is this learner for regression tasks?

id [character(1)]

Id string of learner.

pack [char]

Package(s) required for underlying learner.

doubles [logical(1)]

Can real-valued inputs be processed?

factors [logical(1)]

Can factor inputs be processed?

missings [logical(1)]

Can missing values be processed?

weights [logical(1)]

Can case weights be used?

par.vals [named list]

List of set hyperparameters.

par.set [named list]

Named list of Parameter description objects for all possible hyperparameters.

Further getters for classifiers.

oneclass [logical(1)]

Can oneclass problems be handled?

twoclass [logical(1)]

Can twoclass problems be handled?

multiclass [logical(1)]

Can multiclass problems be handled?

prob [logical(1)]

Can probabilities be predicted?

predict.type [character]

What should be predicted: “response” or “prob”

Setters: setId, setHyperPars, setPredictType

Methods

getParameterSet

getHyperPars

getProperty

setProperty

setProperties

makeWrappedModel

resample

setHyperPars

setId

setPredictType

train