LearnerParameter-class	R Documentation

Description class for a hyperparameter.

Description

Description class for a hyperparameter.

Slots

has.default:

(logical) Was a default value provided?

pass.default:

(logical) Should the default value be always passed down to the underlying function?

default:

(ANY) Default value for this parameter.

when:

(character) Specifies when parameter is used. Possible entries are “train”, “predict” or “both”.

requires:

(expression) R expression over the other parameters to define requirements which make this parameter effective.

Extends

Parameter