Measure-class	R Documentation

Class for performance measures.

Description

A measure object encapsulates a function to evaluate the performance of a prediction. Information about already implemented measures can be obtained here: measures. User-defined measures can be created with makeMeasure.

Details

A learner is trained on a a training set d1, results in a model m, predicts another set d2 (which may be a different one or the training set), resulting in the prediction. The performance measure can now be defined using all of the information of the original task, the fitted model and the prediction.

Slots

id:

(character) Name of the measure.

minimize:

(logical) Should the performance measure be minimized?

classif:

(logical) Is the measure applicable for classification?

regr:

(logical) Is the measure applicable for regression?

only.binary:

(logical) Can the measure only be used for binary classification?

req.pred:

(logical) Does the calculation require a prediction object?

req.task:

(logical) Does the calculation require the task?

req.model:

(logical) Does the calculation require the model?

allowed.pred.types:

(character) Which prediction types are allowed for this measure? Subset of “response”,“prob”.

fun:

(function) Calculation function.

extra.args:

(list) Extra parameters for calculation.

aggr:

(Aggregation) Associated aggregation function.

Methods

performance

setAggregation

See Also

measures, makeMeasure