Parameter-class	R Documentation

Description class for an optimization parameter.

Description

Description class for an optimization parameter.

Details

Subclasses: LearnerParameter

Slots

id:

(character) Name of parameter.

type:

(character) Data type of parameter. Possible types are 'numeric', 'numericvector', 'integer', 'ordered', 'discrete', 'function', 'untyped'.

constraints:

(list) Contains further information of simple contraints. For 'numeric' and 'integer' it contains box constraints 'lower' and 'upper', both numeric(1). For 'discrete' and 'ordered' it contains possible values 'vals' as a list, for 'ordered' the order of the list reflects the implied ordering. For 'numericvector' it contains box constraints 'lower' and 'upper', both numeric(dimension of vector).

trafo:

(function) When values for the parameter are generated in any way, this function will always be applied directly aftwerwards. Function must accept a parameter value as the first argument and return a transformed one.

Methods

lower

upper

values

isFeasible

See Also

makeParameterSet