Prediction-class	R Documentation

Prediction.

Description

Result from predict. Use as.data.frame to access all information in a convenient format.

Details

Prediction-class: Getter.

predict.type [character(1)]

Type set in predict function: "response" or "prob".

id [numeric]

Vector of index numbers of predicted cases from the task.

response [numeric | factor]

Predicted response values.

truth [numeric | factor]

True target values.

prob [numeric | matrix] Optional parameters: class

Predicted probabilities. For binary class: Only the probabilities for the positive class are returned.

threshold [numeric]

Threshold set in predict function.

Methods

as.data.frame

signature( = "Prediction"): Convert to data.frame

getProb

as.ROCR.prediction

getConfMatrix

performance

setThreshold

See Also

performance