ResampleDesc-class	R Documentation

Base class for description of resampling algorithms.

Description

Base class for description of resampling algorithms. A description of a resampling algorithm contains all necessary information to create a ResampleInstance, when given the size of the data set. For construction simply use the factory method makeResampleDesc.

Slots

instance.class:

(character) S4 class name of the corresponding ResampleInstance.

id:

(character) Name of resampling strategy.

iters:

(integer) Number of iterations. Note that this is always the complete number of generated train/test sets, so for a 10 times repeated 5fold cross-validation it would be 50.

predict:

(character) What to predict during resampling: 'train', 'test' or 'both' sets.

Methods

makeResampleInstance

makeResampleInstance

iters