ResampleInstance-class	R Documentation

Resampling instance.

Description

Base class for specific resampling draws like cross-validation or bootstrapping. This class encapsulates training and test sets generated from the data set for a number of iterations. It mainly stores a set of integer vectors indicating the training and test examples for each iteration. Don't create objects from this class directly but use the corresponding subclasses. For construction simply use the factory method makeResampleInstance.

Slots

desc:

(ResampleDesc) [ResampleDesc] Resample description that was used to create the instance.

size:

(integer) [integer(1)] Number of observations.

train.inds:

(list) [list of integer] List of of training indices for all iterations.

test.inds:

(list) [list of integer] List of of test indices for all iterations.

group:

(factor) [factor] Optional grouping of resampling iterations. This encodes whether specfic iterations 'belong together' (e.g. repeated CV), and it can later be used to aggregate performance values accordingly. Default is 'factor()'.

Methods

iters

resample

Note

If you want to add another resampling strategy, have a look at the web documentation.

See Also

ResampleDesc, makeResampleInstance, resample