ResamplePrediction-class	R Documentation

Prediction from resampling.

Description

Contains predictions from resampling, returned (among other stuff) by function resample. Can basically be used as its super class. The main differences are: (a) The internal data.frame (slot df) contains an additional column id, specifying the iteration of the resampling strategy. (b) The object can be converted into a list of Prediction objects by using as.list on it, one object for each resampling iteration.

Slots

instance:

(ResampleInstance) Resampling instance that was used to generate the prediction.

Extends

Prediction

Methods

as.list

signature( = "ResamplePrediction"): Converts object to a list of a Prediction objects - one for each iteration.

as.ROCR.prediction

See Also

resample, predict