WrappedModel-class	R Documentation

Induced model of learner.

Description

Result from train. It internally stores the underlying fitted model, the IDs of the subset used for training, variables used for training and information about second-level optimization like tuned hyperparameters or selected variables.

Details

Getter.
Note that all getters of TaskDesc can also be used.

learner [Learner]

Learner that was used to fit the model.

learner model [any]

Underlying model from used R package.

subset [integer]

Subset used for training.

fail [NULL | string]

Generally NULL but if the training failed, the error message of the underlying train function.

Methods

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

performance

predict

signature(object = "WrappedModel"): Predict the target variable of new data using a fitted model. If the type is set to "prob" probabilities will be stored in the resulting object. The resulting class labels are the classes with the maximum values.

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner

predictLearner