adaboost-class(mlr)	R Documentation

adaboost

Description

Wrapped learner for Adaboost.M1 from package adabag for classification problems.

Details

Common hyperparameters:

minsplit

Minimum number of observations that must exist in a node in order for a split to be attempted.

cp

Complexity parameter. Any split that does not decrease the overall lack of fit by a factor of cp is not attempted.

maxdepth

Maximum depth of any node of the final tree, with the root node counted as depth 0. Defaults to the number of classes.

Extends

wrapped.learner.classif

Methods

See Also

adaboost.M1

[Package mlr version 0.3.174 Index]