bench.exp	R Documentation

Benchmark experiment for multiple learners and tasks.

Description

Complete benchmark experiment to compare different learning algorithms across one or more tasks w.r.t. a given resampling strategy. Experiments are paired, meaning always the same training / test sets are used for the different learners.

Usage

bench.exp(learners, tasks, resampling, measures, conf.mats=TRUE, predictions=FALSE, models=FALSE, paths=FALSE)

Arguments

learners
[string | learner | list of the previous two]
Defines the learning algorithms which should be compared.

tasks
[learn.task | list of the previous]
Defines the tasks.

resampling
[resampling desc | resampling instance | list of the previous two]
Defines the resampling strategies for the tasks.

measures
[see measures] Performance measures.

conf.mats
[logical]
Should confusion matrices be stored? Default is TRUE. Ignored for regression.

predictions
[logical]
Should all predictions be stored? Default is FALSE.

models
[logical]
Should all fitted models be stored? Default is FALSE.

paths
[logical]
Should the optimization paths be stored? Default is FALSE.

Value

bench.result.

Note

You can also get automatic, internal tuning by using make.tune.wrapper with your learner.

See Also

make.tune.wrapper