bench.result-class	R Documentation

bench-result

Description

Container for the results of a benchmark experiment.

Details

bench.result-class: Getter.

The "perf" getter is probably the most common one, it returns a list of 3 dim. arrays of performance values for every data set. The dimension are: learners, resampling iterations and measures. You can reduce the list or the contained array by using the optional arguments "task", "learner", "measure", "iter" and "aggr". "task" and "learners" must be set to a char vector of IDs repectively, "measure" to names of recorded performance measures in the experiment, "iter" to an integer vector of selected resampling interations. The default for these is to select everything. "aggr" can be used to aggregate the results accross the resampling interations (see aggregations). The default is not to do any aggregation. You can also set "aggr" to "resampling" which does the default aggregation of the used resampling stratgegy. 'drop' is by default TRUE, which means that the structures are simplified as much as possible, if you don't want this set 'drop' to FALSE.

The following getters all return list of lists of objects: prediction, conf.mat The first list iterates the tasks, the second one the learners, both are named by respective IDs. You can reduce these lists by using the optional arguments 'task' and 'learner'. 'drop' is by default TRUE, which means that the list structures are simplified as much as possible, if you don't want this set 'drop' to FALSE.

The following getters all return list of lists of lists: opt.result, opt.par, opt.perf, opt.path, tuned.par, sel.var The first list iterates the tasks, the second one the learners, both are named by respective IDs, the third list iterates the resampling iterations. You can reduce these lists by using the optional arguments 'task' and 'learner'. 'drop' is by default TRUE, which means that the list structures are simplified as much as possible, if you don't want this set 'drop' to FALSE.

learners [character]

IDs of learners used in experiment.

tasks [character]

IDs of tasks used in experiment.

measures [character]

Names of measures recorded in experiment.

iters [numeric]

Named numerical vector which lists the number of iterations for every task. Names are IDs of task.

prediction [see above]

List of list of predictions for every task/learner.

conf.mat [see above]

List of list of confusion matrices for every task/learner.

opt.result [see above]

List of list of list of opt.result for every task/learner/iteration. Entry is NULL if no optimization was done.

opt.perf [see above]

List of list of list of performance vectors of optimal settings for every task/learner/iteration. Note that this performance refers to the inner resampling! Entry is NULL if no optimization was done.

opt.par [see above]

List of list of list of optimal settings for every task/learner/iteration. Entry is NULL if no optimization was done.

opt.path [see above]

List of list of list of optimization paths for every task/learner/iteration. Entry is NULL if no optimization was done.

tuned.par [see above]

List of list of list of optimal hyperparameters for every task/learner/iteration. Entry is NULL if no tuning was done. Basically a different name for "opt.par".

sel.var [see above]

List of list of list of optimal features for every task/learner/iteration. Entry is NULL if no feature selection was done.. Basically a different name for "opt.par".

perf [see above]

List of 3 dim. arrays of performance values for every data set.

Extends

object

Methods

[

to.string

See Also

bench.exp