blackboost.regr-class(mlr)	R Documentation

blackboost.regr

Description

Wrapped learner for Boosting with Regression Trees from package mboost for regression problems.

Details

Common hyperparameters:

mstop

Integer, giving the number of initial boosting iterations.

nu

Double (between 0 and 1), defining the step size or shrinkage parameter.

constraint

Logical, indicating whether the working responses should be restricted to (-1, +1).

risk

Character, indicating how the empirical risk should be computed for each boosting iteration.

center

Logical, indicating if the numerical covariates should be mean centered before fitting.

Extends

wrapped.learner.regr

Methods

See Also

blackboost, boost_control

[Package mlr version 0.3.174 Index]