bs.instance-class(mlr)	R Documentation

bs.instance

Description

A bs.instance object is mainly a set of integer vectors generated by bootstrapping. They indicate the training part of data for different instances of bs.

Extends

resample.instance

Methods

initialize

signature( = "bs.instance"): Constructor.

Note

Normally you don't want to use the S4 constructor, but instead make.bs.instance.

[Package mlr version 0.3.180 Index]