classif.task-class(mlr)	R Documentation

classif.task

Description

General description object for a classification experiment. Instantiate it by using its factory method.

Slots

type:

(character) "class" if you generally want to predict classes or "prob" for probabilities. Default is "class"

Extends

learn.task

Methods

predict

signature(object = "classif.task"): Predicts the target classes of a new data set based on an already fitted wrapped.model of a classifcation task.

initialize

signature( = "classif.task"): Constructor.

[

signature( = "classif.task"): Getter.

See Also

make.regr.task

[Package mlr version 0.3.174 Index]