cv.instance-class(mlr)	R Documentation

cv.instance

Description

A cv.instance object is mainly a set of integer vectors generated by cross-validation. They indicate the training part of data for different instances of cv.

Extends

resample.instance

Methods

initialize

signature( = "cv.instance"): Constructor.

Note

Normally you don't want to use the S4 constructor, but instead make.cv.instance.

[Package mlr version 0.3.174 Index]