data.desc-class(mlr)	R Documentation

data.desc

Description

Since not all classifiers can deal with all kind of data, the properties of the data are compared with the supported features of the learner when a learn.task is generated. A data.desc object contains a description of these data properties.

Slots

target.col:

(integer) Column index of the response

is.classification:

(logical) Is the target variable categorical?

class.nr:

(integer) Does the dataset have missing values?

has.missing:

(logical) Does the dataset have missing values?

numerics:

(integer) Does the dataset have numeric variables?

integers:

(integer) Does the dataset have integer variables?

factors:

(integer) Does the dataset have factor variables?

characters:

(integer) Does the dataset have character variables?

obs:

(integer) Number of examples in dataset.

Methods

initialize

print

as.character

print

show

[Package mlr version 0.3.86 Index]