errorhandler.setup	R Documentation

Errorhandler setup.

Description

Sets up the error handling system of mlr.

Usage

errorhandler.setup(on.learner.error="warn", on.par.without.desc="warn",
    on.convert.var="warn")

Arguments

on.learner.error
[string]
What should happen if an error in an underlying learning algorithm is caught: 'stop', 'warn' or 'quiet'.
'stop': R exception is generated.
'warn': Model learner.failure will be created, which predicts only NAs and a warning will be generated.
'quiet': Same as 'warn' but withou the warning.
Default is 'warn'.

on.par.without.desc
[string]
What should happen if a parameter of a learner is set to a value, but no parameter description object exists, indicating a possibly wrong name: 'stop', 'warn' or quiet.
'stop': R exception is generated.
'warn': Warning, but parameter is still passed along to learner.
'quiet': Same as 'warn' but without the warning.
Default is 'warn'.

on.convert.var
[string]
What should happen if an a variable is converted during the creation of a learn.task: 'warn' or 'quiet'.
Currently, integers are converted to numerics, characters to factors and for classification, target levels without corresponding observation are dropped.
'warn': Warning on conversion.
'quiet': Conversion done quietly.
Default is 'warn'.

Value

NULL.