get.learners	R Documentation

Find matching learning algorithms.

Description

Returns the names of learning algorithms which have specific characteristics, e.g. whether it supports missing values, misclassification costs, case weights,... or which are are able to solve a given learn.task.

Arguments

x
[string | learn.task]
Type of the learning algorithm, either "classif" or "regr" or task to solve

numerics
[boolean]
Supports numeric inputs? Pass only when x is a string.

factors
[boolean]
Supports factor inputs? Pass only when x is a string.

characters
[boolean]
Supports character inputs? Pass only when x is a string.

missings
[boolean]
Supports missing values? Pass only when x is a string.

multiclass
[boolean]
Supports multiclass problems? Pass only when x is a string.

weights
[boolean]
Supports case weights? Pass only when x is a string.

probs
[boolean]
Can predict probabilities?

decision
[boolean]
Supports decision values?

costs
[boolean]
Supports non-standard misclassification costs?

Details

get.learners-methods: The default of all boolean parameters is NA, meaning: property is not required, don't care.