getData	R Documentation

Extract data in task.

Description

Extract data in task. Useful in trainLearner when you add a learning machine to the package.

Usage

getData(task, subset, vars, target.extra=FALSE, class.as="factor")

Arguments

task
[LearnTask]
Learning task.

subset
[integer]
Selected cases. Default is all cases.

vars
[character]
Selected inputs. Default is all input variables.

target.extra
[logical(1)]
Should target vector be returned separately? If not, a single data.frame including the target is returned, otherwise a list with the input data.frame and an extra vector for the targets. Default is FALSE.

class.as
[character(1)]
Should target classes be recoded? Only for binary classification. Possible are “factor” (do nothing), “01”, and “-1+1”. In the two latter cases the target vector, which is usually a factor, is converted into a numeric vector. The positive class is coded as +1 and the negative class either as 0 or -1. Default is “factor”.

Value

Either a data.frame or a list with data.frame data and vector target.