getLearners	R Documentation

Find matching learning algorithms.

Description

Returns the names of learning algorithms which have specific characteristics, e.g. whether it supports missing values, case weights,... or which are are able to solve a given LearnTask.

Arguments

x
[string | LearnTask]
Type of the learning algorithm, either "classif" or "regr" or task to solve

doubles
[logical(1)]
Supports real-valued inputs? Pass only when x is a string.

factors
[logical(1)]
Supports factor inputs? Pass only when x is a string.

characters
[logical(1)]
Supports character inputs? Pass only when x is a string.

missings
[logical(1)]
Supports missing values? Pass only when x is a string.

multiclass
[logical(1)]
Supports multiclass problems? Pass only when x is a string.

weights
[logical(1)]
Supports case weights? Pass only when x is a string.

probs
[logical(1)]
Can predict probabilities?

Details

getLearners-methods: The default of all boolean parameters is NA, meaning: property is not required, don't care.