getProb	R Documentation

Get probabilities for some classes.

Description

Get probabilities for some classes.

Arguments

pred
[Prediction] Prediction object.

class
[character] Names of classes. Default is either all classes for multi-class problems or the positive class for binary classification.

Value

getProb-methods: Data.frame with numerical columns or a numerical vector if length of class is 1. Order of columns is defined by class.