getter,resample.instance-method(mlr)	R Documentation

Getter for resample.instance

Description

Getter.

Arguments

x 	 resample.instance object

i 	 [character]

data.size

The size of the dataframe.

name

The name of the resample description object, i.e. the type of resampling.

iters

The number of resampling iterations.

train.inds

If j is a single integer, the vector of training indices for the jth iteration is returned. If j is an integer vector, the list of training indices for the given iterations is returned. If j is missing, all indices are returned.

test.inds

If j is a single integer, the vector of test indices for the jth iteration is returned. If j is an integer vector, the list of test indices for the given iterations is returned. If j is missing, all indices are returned.

j 	 [integer]
See above, i == "train.inds" or i == "test.inds".

[Package mlr version 0.3.170 Index]