grid.control	R Documentation

grid.control-class

Description

Control structure for grid search tuning.

Arguments

minimize
[logical]
Minimize performance measure? Default is TRUE.

tune.threshold
[logical]
Perform empirical thresholding? Default is FALSE. Only supported for binary classification and you have to set predict.type to "prob" for this in make.learner.

thresholds
[numeric]
Number of thresholds to try in tuning. Predicted probabilities are sorted and divided into groups of equal size. Default is 10.

path
[boolean]
Should optimization path be saved?

ranges
[list]
A list of named vectors/lists of possible values for each hyperparameter. You can also pass a list of such ranges by using [combine.ranges] in the rare case when it does not make sense to search a complete cross-product of range values.

scale
[function]
A function to scale the hyperparameters. E.g. maybe you want to optimize in some log-space. Has to take a vector and return a scaled one. Default is identity function.

Value

grid.control-methods: Control structure for tuning.

Extends

tune.control