initialize,resample.instance-method(mlr)	R Documentation

This is mainly for internal use you only need to use this when you extend resample...

Description

This is mainly for internal use, you only need to use this, when you extend resample.instance to add another resampling strategy!

Arguments

desc 	 [resample.desc]
resample.desc. Describes the resampling strategy.

size 	 [integer]
Size of the data set to resample from.

inds 	 [list of integer vectors]
Indices of the trainings sets.

[Package mlr version 0.3.180 Index]