kernlab.svm.classif-class(mlr)	R Documentation

kernlab.svm.classif

Description

Wrapped learner for Support Vector Machines from package kernlab for classification problems.

Details

General hyperparameters:

C

Cost of constraints violation (default: 1), this is the 'C'-constant of the regularization term in the Lagrange formulation.

Kernel hyperparameters, sorted by kernel type:

rbfdot

sigma: inverse kernel width

laplacedot

sigma: inverse kernel width

polydot

degree: degree of the polynomial,
scale: scaling parameter of the polynomial,
offset: offset used in the polynomial

tanhdot

scale: scaling parameter of the tangent kernel,
offset: offset used in the hyperbolic tangent kernel

besseldot

sigma: inverse kernel width,
ordner: order of the Bessel function,
degree: degree of the Bessel function

anovadot

sigma: inverse kernel width,
degree: degree of the ANOVA kernel function

stringdot

length: length of the strings considered,
lambda: the decay factor,
normalized: logical parameter determining if the kernel evaluations should be normalized.

The kernel type and the hyperparameters are specified in parset.

Extends

wrapped.learner.classif

Methods

initialize

signature( = "kernlab.svm.classif"): Constructor.

train.learner

signature(wrapped.learner = "kernlab.svm.classif", formula = "formula", data = "data.frame", weights = "numeric", parset = "list"): Overwritten, to allow direct passing of kernel hyperparameters. Besides that, simply delegates to super method.

See Also

ksvm, dots

[Package mlr version 0.3.180 Index]