learner-class	R Documentation

Base class for inducers.

Description

Abstract base class for learning algorithms.

Details

learner-class: Getter.
Note that all getters of learner.desc can also be used.

is.classif [boolean]

Is this learner for classification tasks?

is.regr [boolean]

Is this learner for regression tasks?

id [string]

Id string of learner.

label [string]

Label string of learner.

pack [string]

Package were underlying learner is implemented.

desc [learner.desc]

Properties object to describe functionality of the learner.

par.vals [list]

List of fixed hyperparameters and respective values for this learner.

par.vals.name [character]

Names of currently fixed hyperparameters.

par.descs [list]

Named list of par.desc description objects for all possible hyperparameters for this learner.

par.descs.name [character]

Names of all hyperparameters for which description objects exist.

par.descs.when [character]

Named character vector. Specifies when a cetrain hyperparameter is used. Possible entries are 'train', 'predict' or 'both'.

predict.type [character]

What should be predicted: 'response', 'prob' or 'decision'.

predict.threshold [character]

Threshold to produce class labels if type is not "response".

Extends

object

Methods

[

signature( = "learner"): Getter.

to.string

resample.fit

set.hyper.pars

set.id

set.label

train