learner.props-class(mlr)	R Documentation

Since not all classifiers can deal with all kind of data properties of the data are compared...

Description

Since not all classifiers can deal with all kind of data properties of the data are compared with the possibilities of the classifier when a learn.task is generated. A classif.props object describes such classifier possibilities by logical values.

Slots

supports.missing:

(logical) [logical] Can the classifier deal with missing values?

supports.numerics:

(logical) [logical] Can the classifier deal with numeric variables?

supports.factors:

(logical) [logical] Can the classifier deal with factor variables?

supports.characters:

(logical) [logical] Can the classifier deal with character variables?

supports.weights:

(logical) [logical] Can the classifier return class probabilities when predicting?

Methods

show

signature( = "learner.props"): Displays the object

print

signature( = "learner.props"): Prints the object

Note

When a learn.task gets printed its classifier description is shown on the console.

[Package mlr version 0.3.115 Index]