learners	R Documentation

Implemented inducers.

Description

List of supported learning algorithms.

Details

classif.ada
Boosting from ada package: ada

classif.adaboost.M1
Boosting from adabag package: adaboost.M1

classif.blackboost
Gradient boosting with regression trees from mboost package: blackboost

classif.ctree
Conditional Inference Trees from party package: ctree

classif.gbm
Gradient boosting machine from gbm package: gbm

classif.glmboost
Boosting for GLMs from mbboost package: glmboost

classif.grplasso
Logistic Regression with Group Lasso from grplasso package: grplasso

classif.j48
J48 Decision Trees from RWeka package: J48

classif.JRip
Propositional Rule Learner from RWeka package: JRip

classif.kknn
k-Nearest Neighbor from kknn package: kknn

classif.ksvm
Support Vector Machines from kernlab package: ksvm
Note that kernel parameters have to be passed directly and not by using the kpar list in ksvm.

classif.lda
Linear Discriminant Analysis from MASS package: lda

classif.loclda
Local LDA from klaR package: loclda

classif.logreg
Logistic Regression from stats package: glm

classif.lssvm
Least Squares Support Vector Machine from kernlab package: lssvm

classif.lvq1
Learning Vector Quantization from class package: lvq1

classif.mda
Mixture Discriminant Analysis from mda package: mda

classif.multinom
Multinomial Regression from nnet package: multinom

classif.naiveBayes
Naive Bayes from e1071 package: naiveBayes

classif.nnet
Neural Network from nnet package: nnet

classif.OneR
1-R classifier from RWeka package: OneR

classif.PART
PART decision lists from RWeka package: PART

classif.penalizedSVM
Support Vector Machines with L1 penalty from penalizedSVM package: penalizedSVM

classif.qda
Quadratic Discriminant Analysis from MASS package: qda

classif.randomForest
Random Forest from randomForest package: randomForest

classif.rda
Regularized Discriminant Analysis from klaR package: rda

classif.rpart
Decision Tree from rpart package: rpart

classif.svm
Support Vector Machines (libsvm) from e1071 package: svm

regr.blackboost
Gradient boosting with regression trees from mboost package: blackboost

regr.earth
Multivariate Adaptive Regression Splines from earth package: earth

regr.gbm
Gradient boosting machine from gbm package: gbm

regr.kknn
K-Nearest-Neighbor regression from kknn package: kknn

regr.km
Kriging from DiceKriging package: km

regr.ksvm
Support Vector Machines from kernlab package: ksvm
Note that kernel parameters have to be passed directly and not by using the kpar list in ksvm.

regr.lasso
Lasso regression from penalized package: penalized

regr.lm
Simple linear regression from stats package: lm

regr.nnet
Neural Network from nnet package: nnet

regr.randomForest
Random Forest from randomForest package: randomForest

regr.ridge
Ridge regression from penalized package: penalized

regr.rpart
Decision Tree from rpart package: rpart

regr.rsm
Response surface regression from rsm package: rsm
Note that you select the order of the regression by using modelfun="FO" (first order), "TWI" (two-way interactions, this is with 1st oder terms!) and "SO" (full second order).

regr.rvm
Relevance Vector Machine from rpart kernlab: rvm
Note that kernel parameters have to be passed directly and not by using the kpar list in rvm.