logreg-class(mlr)	R Documentation

logreg

Description

Wrapped learner for Logistic Regression from package stats for classification problems.

Details

Common hyperparameters:

start

Starting values for the parameters in the linear predictor.

etastart

Starting values for the linear predictor.

mustart

Starting values for the vector of means.

Extends

wrapped.learner.classif

Methods

See Also

glm

[Package mlr version 0.3.180 Index]