losses	R Documentation

Loss functions.

Description

Losses can always be passed as a single string (name of a single loss), a character vector of multiple names of losses or a list containing string names of losses and your own loss functions as function objects. The latter ones should be named list elements.

Details

Classification:

zero-one
Zero-one loss

costs
Misclassification costs according to cost matrix

Regression:

squared
Squared error

abs
Absolute error

residual
Signed residual error