make.bs.instance(mlr)	R Documentation

make.bs.instance

Description

make.bs.instance generates a bs.instance object, which encapsulates the generated indices of training and test sets.

Usage

make.bs.instance(size, iters)

Arguments

size 	 [integer]
Size of the data set to resample.

iters 	 [integer]
Number of generated subsets / resampling iterations.

Value

A bs.instance object, which encapsulates the generated indices of training and test sets.

See Also

resample.fit, bs.instance

Examples

data(iris)
rin <- make.bs.instance(size=nrow(iris), iters=10)

[Package mlr version 0.3.180 Index]