make.cv.instance(mlr)	R Documentation

make.cv.instance

Description

make.cv.instance generates a cv.instance object, which encapsulates the generated indices of training and test sets.

Usage

make.cv.instance(size, iters)

Arguments

size 	 [integer]
Size of the data set to resample.

iters 	 [integer]
Number of generated subsets / resampling iterations.

Value

A cv.instance object, which encapsulates the generated indices of training and test sets.

See Also

resample.fit, cv.instance

Examples

data(iris)
rin <- make.cv.instance(size=nrow(iris), iters=10)

[Package mlr version 0.3.180 Index]