make.filter.wrapper	R Documentation

Fuse learner with filter method.

Description

Fuses a base learner with a filter method. Creates a learner object, which can be used like any other learner object. Currently only filtering to manually selected variables is supported.

Usage

make.filter.wrapper(learner, id=as.character(NA), label=as.character(NA), vars)

Arguments

learner
[learner or string]
Learning algorithm. See learners.

id
[string]
Id for resulting learner object. If missing, id of "learner" argument is used.

label
[string]
Label for resulting learner object. If missing, label of "learner" argument is used.

vars
[character]
Selected variables.

Value

learner.