make.learner	R Documentation

Create learner object.

Description

Create learner object.

Usage

make.learner(class, id, label, predict.type="response",
    predict.threshold=numeric(0), ..., par.vals=list())

Arguments

class
[string]
Class of learner to create.

id
[string]
Id string for object. Used to select the object from a named list, etc.

label
[string]
Label string for object. Used in plots, etc.

predict.type
[string]
Classification: "response" | "prob" | "decision", specifying the type to predict. Default is "response". "decision" is experimental. Ignored for regression.

predict.threshold
[numeric]
Threshold to produce class labels if type is not "response". Currently only supported for binary classification and type="prob", where it represents the required predicted probability for the positive class, so that a positive class is predicted as "response". Default is 0.5 for type="prob". Ignored for regression.

...
[any]
Optional named (hyper)parameters. Alternatively, you can pass via the "par.vals" argument.

par.vals
[list]
Optional list of named (hyper)parameters. Alternatively, you can pass via the ... argument.

Value

learner.