make.preproc.wrapper	R Documentation

Fuse learner with preprocessing.

Description

Fuses a base learner with a preprocessing method. Creates a learner object, which can be used like any other learner object, but which internally preprocesses the data as requested. If the train or predict function is called on it, the preprocessing is always invoked before.

Usage

make.preproc.wrapper(learner, id=as.character(NA), label=as.character(NA), fun, ...)

Arguments

learner
[learner or string]
Learning algorithm. See learners.

id
[string]
Id for resulting learner object. If missing, id of "learner" argument is used.

label
[string]
Label for resulting learner object. If missing, label of "learner" argument is used.

fun
[function]
Function to preprocess a data.frame. First argument must be called 'data', which will be preprocessed and subsequently returned.

...
[any]
Optional parameters to control the preprocessing. Passed to fun.

Value

learner.