make.subsample.instance(mlr)	R Documentation

make.subsample.instance

Description

make.subsample.instance generates a subsample.instance object, which encapsulates the generated indices of training and test sets.

Usage

make.subsample.instance(size, split, iters)

Arguments

size 	 [integer]
Size of the data set to resample.

split 	 [numeric]
Proportion of data used for training set.

iters 	 [integer]
Number of generated subsets / resampling iterations.

Value

A subsample.instance object, which encapsulates the generated indices of training and test sets.

See Also

resample.fit, subsample.instance

Examples

data(iris)
# split is the training set percentage
rin <- make.subsample.instance(size=nrow(iris), split=2/3, iters=10)
# holdout
rin <- make.subsample.instance(size=nrow(iris), split=2/3, iters=1)

[Package mlr version 0.3.86 Index]