make.tune.wrapper	R Documentation

Fuse learner with tuning.

Description

Fuses a base learner with a search strategy to select its hyperparameters. Creates a learner object, which can be used like any other learner object, but which internally uses tune. If the train function is called on it, the search strategy and resampling are invoked to select an optimal set of hyperparameter values. Finally, a model is fitted on the complete training data with these optimal hyperparameters and returned.

Usage

make.tune.wrapper(learner, id=as.character(NA), label=as.character(NA), resampling,
    control, measures, aggr)

Arguments

learner
[learner or string]
Learning algorithm. See learners.

id
[string]
Id for resulting learner object. If missing, id of "learner" argument is used.

label
[string]
Label for resulting learner object. If missing, label of "learner" argument is used.

resampling
[resample.instance] or [resample.desc]
Resampling strategy to evaluate points in hyperparameter space.

control
Control object for search method. Also selects the optimization algorithm for tuning.

measures
[see measures] Performance measures.

aggr
[see aggregations] Aggregation functions.

Value

learner.

See Also

tune, grid.control, optim.control, cmaes.control