makeDiscreteParameter	R Documentation

Discrete variable for optimization.

Description

Discrete variable for optimization.

Usage

makeDiscreteParameter(id, vals)

Arguments

id
[character(1)] Name of parameter.

vals
[list | vector]
Possible values. You are allowed to pass a list of complex R values, w which are used as discrete choices during optimization. If you do that, make sure that the elements are uniquely named, so that the names can be used as internal represenatation for the choice.

Value

Parameter