makeFunctionLearnerParameter	R Documentation

Function parameter for a learner.

Description

Function parameter for a learner.

Usage

makeFunctionLearnerParameter(id, default, pass.default=FALSE, when="train", requires=expression())

Arguments

id
[character(1)] Name of parameter.

default
[any] Default value used in learner. If this argument is missing, it means no default value is available.

pass.default
[logical(1)] Should the default value be always passed to the learner? Default is FALSE.

when
[character(1)] When is the parameter used in the corresponding learner, either 'train', 'predict' or 'both'. Default is 'train'.

requires
[expression] R expression over the other parameters to define requirements which make this parameter effective.

Value

Parameter