makeIntegerVectorLearnerParameter	R Documentation

Integer vector parameter for a learner.

Description

Integer vector parameter for a learner.

Usage

makeIntegerVectorLearnerParameter(id, dim=NA, lower=-Inf, upper=Inf, default, pass.default=FALSE,
    when="train", requires=expression())

Arguments

id
[character(1)] Name of parameter.

dim
[integer(1)] Dimension of vector. Can be NA, which means dimension in unknown. Default is NA.

lower
[integer(1)]
Lower bound. Default is -.Machine$integer.max.

upper
[integer(1)]
Upper bound. Default is Machine$integer.max.

default
[numeric(1)] Default value used in learner. If this argument is missing, it means no default value is available.

pass.default
[logical(1)] Should the default value be always passed to the learner? Default is FALSE.

when
[character(1)] When is the parameter used in the corresponding learner, either 'train', 'predict' or 'both'. Default is 'train'.

requires
[expression] R expression over the other parameters to define requirements which make this parameter effective.

Value

Parameter