makeLearner	R Documentation

Create learner object.

Description

Create learner object.

Usage

makeLearner(class, id, predict.type="response", ..., par.vals=list())

Arguments

class
[character(1)]
Class of learner to create. By convention, all classification learners start with “classif.” and all regression learners with “regr.”. A list of all learners is available on the learners help page.

id
[character(1)]
Id string for object. Used to select the object from a named list, etc.

predict.type
[character(1)]
Classification: “response” or “prob”, specifying the type of prediction returned. The default is “response”. For regression, this argument is ignored and should not be set.

...
[any]
Optional named (hyper)parameters. Alternatively these can be given using the par.vals argument.

par.vals
[list]
Optional list of named (hyper)parameters. The arguments in ... take precedence over values in this list. We strongly encourage you to use one or the other to pass (hyper)parameters to the learner but not both.

Details

To change an object later on look at the setters of the Learner class.

For a classification learner the predict.type can be set to “prob” to predict probabilities and the maximum value selects the label. The threshold used to assign the label can later be changed using the setThreshold function.

Value

Learner.