makeRegrTask	R Documentation

Construct learning task.

Description

Defines a regression task for a given data set. The target variable is converted to a numeric.

Arguments

id
[character(1)]
Id string for object. Used to select the object from a named list, etc. Default is the name of R variable passed to data.

data
[data.frame]
A data frame containing the input and target variables for modeling.

target
[character(1)]
Name of the target variable.

exclude
[character] Names of inputs, which should be discarded, e.g. IDs, etc. Default is zero-length vector.

weights
[numeric]
An optional vector of case weights to be used in the fitting process (if the learner cannot handle weights, they are ignored). Default is not to use weights.

blocking
[factor]
An optional factor of the same length as the number of observations. Observations with the same blocking level "belong together". Specifically, they are either put all in the training or the test set during a resampling iteration.

check.data
[logical(1)] Should sanity of data be checked initially at task creation? You should have good reasons to turn this off...

Value

makeRegrTask-methods: LearnTask.