makeResampleDesc	R Documentation

Construct resampling description.

Description

Generates a description object for a resampling strategy.

Arguments

method
[character(1)]
'CV' for cross-validation, 'LOO' for leave-one-out, 'StratCV' for stratified cross-validation, 'RepCV' for repeated cross-validation,
'BS' for out-of-bag bootstrap, 'Subsample' for subsampling, 'Holdout' for holdout.

iters
[integer]
Number of resampling iterations. Ignored for 'Holdout'. Default is 10.

predict
[character]
What to predict during resampling: 'train', 'test' or 'both' sets. Default is 'test'.

...
[any]
Further parameters for strategies.
split [numeric(1)]: Proportion of training cases for 'Holdout' and 'Subsample' from between 0 and 1. Default is 2/3.
reps [integer(1)]: Repeats for 'RepCV'. Here 'iters' = 'folds' x 'reps'. Default is 2.
folds [integer(1)]: Folds in the repeated CV for 'RepCV'. Here 'iters' = 'folds' x 'reps'. Default is 5.

Details

makeResampleDesc-methods: Repeated cross-validation: Use 'RepCV'. Then you have to set the aggregation function for your preferred performance measure to 'testgroup.mean' via setAggregation. B632 bootstrap: Use 'BS' for bootstrap and set predict to 'both'. Then you have to set the aggregation function for your preferred performance measure to 'b632' via setAggregation. B632+ bootstrap: Use 'BS' for bootstrap and set predict to 'both'. Then you have to set the aggregation function for your preferred performance measure to 'b632plus' via setAggregation.

Value

makeResampleDesc-methods: ResampleDesc.