opt.result-class	R Documentation

Optimization result.

Description

Container for results of hyperparameter tuning or variable selection. Contains the obtained optimal parameter vector, its performance values and the optimization path which lead there. It might also optionally contain a wrapped.model, which was fitted by using the optimal parameters.

Details

opt.result-class: Getter.

opt.type [string]

Currently 'tune' or 'varsel'.

par [list | character]

Named list of hyperparameter values or character vector of variables, identified as optimal.

tuned.par [list]

If tuning was performed, best found set of hyperparameters.

sel.vars [character]

If variable selection was performed, best found set of variables.

perf [numeric]

Performance values of 'par'.

path [list | data.frame]. Optional parameters: as.data.frame

Optimization path. Can be converted to a data.frame if as.data.frame is TRUE.

model [wrapped.model]

Model fitted with settings in 'par'. Will be NULL, if fitting was not requested.

learner [wrapped.model]

Learner with settings in 'par'. Currently only supported for hyperparameter tuning.

Extends

object

Methods

[

to.string

See Also

tune, varsel