opt.wrapper-class	R Documentation

Abstract base class to wrap an optimization algorithm around a learner.

Description

Abstract base class to wrap an optimization algorithm around a learner.

Extends

base.wrapper

Methods

[

train.learner