optim.control	R Documentation

optim.control-class

Description

Control structure for tuning with optim (Nelder-Mead, SANN, etc).

Arguments

minimize
[logical]
Minimize performance measure? Default is TRUE.

tune.threshold
[logical]
Perform empirical thresholding? Default is FALSE. Only supported for binary classification and you have to set predict.type to "prob" for this in make.learner.

thresholds
[numeric]
Number of thresholds to try in tuning. Predicted probabilities are sorted and divided into groups of equal size. Default is 10.

path
[boolean]
Should optimization path be saved?

start
[numeric]
Named vector of initial values.

lower
[numeric]
Named vector of lower boundary constraints. Default is -Inf. Only used in "L-BFGS-B".

upper
[numeric]
Named vector of upper boundary constraints. Default is Inf. Only used in "L-BFGS-B".

scale
[function]
A function to scale the hyperparameters. E.g. maybe you want to optimize in some log-space. Has to take a vector and return a scaled one. Default is identity function.

...
Further control parameters passed to the control argument of optim.

Value

optim.control-methods: Control structure for tuning.

Extends

tune.control