Overview

This web page page will provide you with an in-depth introduction on how to use the mlr package for machine learning experiments and optimization in R.

Our focus is on your comprehension of the basic functions and applications.
For technical stuff and special cases you should take a look at the R Documentation.

The approach is step by step and orients itself on the proceeding of a basic data set analysis:

What is your data and the aim of your analysis?
How do you select learning algorithms and set them up?
What about hyperparameters?
How to predict new data?
How to evaluate my model and its predictions?

We present various simple examples from classification and regression during this tutorial to illustrate the main features of the package.

Enjoy reading!