parallel.setup	R Documentation

Parallelization setup.

Description

Defines the technical mode and level of parallelization when executing code.

Usage

parallel.setup(mode="local", parallel.type, cpus, level="resample", ...)

Arguments

mode
[string]
Which parallel mode should be used: "local", "multicore", "snowfall", "sfCluster". Default is "local" without parallel execution.

parallel.type
[string]
Currently this is only used for sfInit (mode="snowfall"), where it is passed to the "type" argument. Default is "MPI", meaning we use "Rmpi".

cpus
[numeric]
Number of requested cpus. Default is mpi.universe.size() for snowfall/MPI, ignored for for sfCluster and 1 otherwise.

level
[string]
What is parallelized / what is a job. "resample": resample.fit is parallelized and a job is train / test. "tune": tune is parallelized and a job is a resampled evaluation of one hyperparameter setting. "varsel": varsel is parallelized and a job is a resampled evaluation of a feature set. "bench": bench.exp is parallelized and a job is completely evaluating one learner on one data set.

...
[any]
Optional parameters, only passed to sfInit currently.

Details

Note that when you parallelize a function, e.g. resample.fit, this also works when the function is internally called by mlr in a more complex function, e.g. bench.exp.