penalized.lasso-class(mlr)	R Documentation

penalized.lasso

Description

Wrapped learner for Lasso Regression from package penalized for regression problems.

Details

Common hyperparameters:

lambda1

Tuning parameter for L1 penalization.

steps

If greater than 1, the algorithm will fit the model for a range of steps lambda1-values, starting from the maximal value down to the value of lambda1 specified.

epsilon

The convergence criterion.

Extends

wrapped.learner.regr

Methods

initialize

signature( = "penalized.lasso"): Constructor.

train.learner

See Also

penalized

[Package mlr version 0.3.100 Index]