pred.learner	R Documentation

pred.learner,classif.ada,wrapped.model,data.frame,character-method

Description

Mainly for internal use. Predicts new data with wrapped.model. You have to implement this method if you want to add another learner to this package.

Arguments

.learner
[\code\linkS4classlearner]
Wrapped learner from this package.

.model
[character]
Model produced by training.

.newdata
[data.frame]
New data to predict.

type
[character]
Specifies the type of predictions - either probability ("prob") or class ("class"). Ignore this if it is not classification or the learner does not support probabilities.

...
[any]
Additional parameters, which need to be passed to the underlying train function.

Value

pred.learner-methods: Model of the underlying learner.