prediction-class	R Documentation

Prediction.

Description

Result from predict. Use as.data.frame to access all information in a convenient format.

Details

prediction-class: Getter.

type [string]

Type set in predict function: "response", "prob", or "decision".

id [numeric]

Vector of index numbers of predicted cases from the task.

response [numeric | factor]

Predicted response values.

truth [numeric | factor]

True target values.

prob [numeric | matrix] Optional parameters: class

Predicted probabilities. For binary class. only the probabilities for the positive class are returned.

decision [matrix]

Predicted decision values.

group [factor]

Grouping of predcitions. Rarely used, if not present NULL is returned.

threshold [numeric]

Threshold set in predict function.

has.groups [boolean]

Is special grouping used in order to aggregate predictions differently?

Extends

object

Methods

as.ROCR.preds

resample.update

performance

[

signature( = "prediction"): Getter.

as.data.frame

signature( = "prediction"): Convert to data.frame

to.string

See Also

performance