randomvarsel.control	R Documentation

randomvarsel.control-class

Description

Control structure for random variable selection.

Arguments

minimize
[logical]
Minimize performance measure? Default is TRUE.

tune.threshold
[logical]
Perform empirical thresholding? Default is FALSE. Only supported for binary classification and you have to set predict.type to "prob" for this in make.learner.

thresholds
[numeric]
Number of thresholds to try in tuning. Predicted probabilities are sorted and divided into groups of equal size. Default is 10.

path
[boolean]
Should optimization path be saved?

maxit
[integer]
Maximal number of variable sets to evaluate. Default is 100.

method
[numeric]
Currently only "binomial" is implemented. Samples variables from a binomial distribution.

prob
[numeric]
Parameter for binomial distribution.

Value

randomvarsel.control-methods: Control structure.

Extends

varsel.control