resample	R Documentation

Fit models according to a resampling strategy.

Description

Given a resampling strategy, which defines sets of training and test indices, fits the selected learner using the training sets and performs predictions for the training/test sets. (This depends on what you selected in the resampling strategy, see parameter predict in makeResampleDesc.) Then performance measures are calculated and aggregated. You are able to return all fitted models (parameter models) or extract specific parts of the models (parameter extract) as returning all of them completely might be memory intensive.

Arguments

learner
[Learner | character]
Learning algorithm.

task
[LearnTask]
Learning task.

resampling
[ResampleDesc or ResampleInstance]
Resampling strategy. If a description is passed, it is instantiated automatically.

measures
[Measure | list of Measure]
Performance measures to evaluate. See measures.

models
[logical(1)]
Should all fitted models be returned? Default is FALSE.

extract
[function(model)]
Function used to extract information from a fitted model during resampling. Is applied to every WrappedModel resulting from calls to train during resampling. Default is to extract nothing.

Details

resample-methods: For construction of the resampling strategies use the factory methods makeResampleDesc and makeResampleInstance.

Value

resample-methods: List of
measures.test [data.frame] Rows correspond to test sets in resampling iterations, columns to performance measures.
measures.train [data.frame] Rows correspond to training sets in resampling iterations, columns to performance measures.
aggr [named numeric] Vector of aggregated performance values. Names are coded like this <measure>.<aggregation>.
pred [ResamplePrediction] Container for all predictions during resampling.
models [list of WrappedModel] List of fitted models or NULL.
extract [list] List of extracted parts from fitted models or NULL.