resample.instance-class	R Documentation

Resampling instance.

Description

Base class for specific resampling draws like cross-validation or bootstrapping. This class encapsulates training and test sets generated from the data set for a number of iterations. It mainly stores a set of integer vectors indicating the training examples for each iteration. Don't create objects from this class directly but use the corresponding subclasses. For construction simply use the factory method make.res.instance.

Details

resample.instance-class: Getter.

size [integer]

Number of observations.

name [character]

The name of the resample description object, i.e. the type of resampling.

iters [integer]

The number of resampling iterations.

train.inds [list | integer] Optional parameter: j

If j is a single integer, the vector of training indices for the jth iteration. Otherwise, the list of indices for iterations. Missing j means list of all indices.

test.inds [list | integer] Optional parameter: j

If j is a single integer, the vector of test indices for the jth iteration. Otherwise, the list of indices for iterations j. Missing j means list of all indices.

Extends

object

Methods

get.train.set

get.test.set

resample.update

resample.done

resample.fit

[

to.string

Note

If you want to add another resampling strategy, have a look at the web documentation.

See Also

resample.desc, make.res.instance, resample.fit