resample.prediction-class	R Documentation

Converts object to a list of lists of normal prediction objects - one for each iteration and group...

Description

Converts object to a list of lists of normal prediction objects - one for each iteration and group

Details

[,resample.prediction-method: Note that in the case of the "prob", "response" and "decision" fields, the results are returned in the order they were used by the resampling strategy and not in the order present in the dataset. This mainly applies to cross-validation were a different order might be expected.

Extends

prediction

Methods

as.ROCR.preds

performance

to.string

[

signature( = "resample.prediction"): Getter

as.list

signature( = "resample.prediction"): Converts object to a list of normal prediction objects - one for each iteration.