resample.result-class(mlr)	R Documentation

resample.result

Description

resample.result contains the results of a resampling process, mainly the list of predicitions and - if saved - the list of fitted models.

Slots

ri.class:

(character) Class of resample.instance

ri.name:

(character) Name of resample.instance

preds:

(list) List of predictions, i.e. list of factors or list of matrices for classification respectivly regression.

models:

(list) List of the fitted models.

Methods

resample.performance

signature(learn.task = "learn.task", resample.instance = "resample.instance", resample.result = "resample.result", measure = "list"): Measures the quality of predictions w.r.t. some loss function for a resampled fit.

[

signature( = "resample.result"): Getter.

See Also

resample.desc, resample.instance, make.cv.instance, make.bs.instance, make.subsample.instance, resample.fit

[Package mlr version 0.3.174 Index]