sequential.control	R Documentation

sequential.control-class

Description

Control structure for sequential variable selection.

Arguments

minimize
[logical]
Minimize performance measure? Default is TRUE.

tune.threshold
[logical]
Perform empirical thresholding? Default is FALSE. Only supported for binary classification and you have to set predict.type to "prob" for this in make.learner.

thresholds
[numeric]
Number of thresholds to try in tuning. Predicted probabilities are sorted and divided into groups of equal size. Default is 10.

path
[boolean]
Should optimization path be saved?

max.vars
[integer]
Maximal number of allowed variables in the final set. Default is max. integer.

method
[character]
Search method. Currently supported are sequential forward search "sfs", sequential backward search "sbs", sequential floating forward search "sffs", sequential floating backward search "sfbs". Default is "sfs".

alpha
[numeric]
sfs, sffs: In a forward step, minimal improvement of performance measure. Can be negative.

beta
[numeric]
sbs, sfbs: In a backward step, minimal improvement of performance measure. Can be negative.

Value

sequential.control-methods: Control structure.

Extends

varsel.control