set.predict.par(mlr)	R Documentation

set.predict.par

Description

Set a parameter for the underlying predict function of a [wrapped.learner] or - for convienience - a [learn.task]. Used to fix (somewhat techical) arguments which stay the same for the whole experiment. You should not have to use this too often.

Usage

set.predict.par(object, ...)set.predict.par(object, ...)

Arguments

object 	 [wrapped.learner]
Wrapping object for the underlying learner.

... 	 Parameters to fix in underlying predict function. Have to be named.

Value

set.predict.par,wrapped.learner-method: Wrapped.learner object with changed paramters for predict function.

[Package mlr version 0.3.180 Index]