setAggregation	R Documentation

Set aggregation function of measure.

Description

Set how this measure will be aggregated after resampling.

Arguments

measure
[Measure]
Performance measure.

aggr
[Aggregation]
Aggregation function.

Details

setAggregation-methods: Possible are: test.mean, test.sd, test.median, test.min, test.max They all calculate the respective statistic of the test set performances.

Value

setAggregation-methods: Measure with changed aggregation behaviour.