setHyperPars	R Documentation

setHyperPars,BaseWrapper,list-method

Description

Set the hyperparameters of learner objects.

Arguments

learner
[Learner]
Learner object.

...
[any]
Optional named (hyper)parameters. Alternatively these can be given using the par.vals argument.

par.vals
[list]
Optional list of named (hyper)parameters. The arguments in ... take precedence over values in this list. We strongly encourage you to use one or the other to pass (hyper)parameters to the learner but not both.

Value

setHyperPars-methods: Learner with changed hyperparameters.

See Also

See getHyperPars for a function to retrieve the currently set hyper parameters. To get a list of all (hyper)parameters of a learner, see the par.set slot of the Learner object.