setThreshold	R Documentation

Set threshold of prediction object.

Description

Set threshold of rediction object (if prediction was for classification). Creates corresponding discrete class response for the newly set threshold. For binary classification: The positive class is predicted if the probability value exceeds the threshold. For multiclass: Probabilities are divided by corresponding thresholds and the class with maximum resulting value is selected. The result of both are equivalent if in the multi-threshold case the labels are greater than 0 and sum to 1.

Arguments

pred
[Prediction]
Prediction object.

threshold
[numeric]
Threshold to produce class labels. Has to be a named vector, where names correspond to class labels. Only if pred is a prediction object resulting from binary classification it can be a single numerical threshold for the positive class.

Value

setThreshold-methods: Prediction with changed threshold and corresponding response.