subsample.instance-class(mlr)	R Documentation

subsample.instance

Description

A subsample.instance object is mainly a set of integer vectors generated by subsampling. They indicate the training part of data for different instances of subsample.

Extends

resample.instance

Methods

initialize

signature( = "subsample.instance"): Constructor.

Note

Normally you don't want to use the S4 constructor, but instead make.subsample.instance.

[Package mlr version 0.3.180 Index]