train.learner	R Documentation

train.learner,classif.ada,character,data.frame,data.desc,task.desc,numeric,matrix-method

Description

Mainly for internal use. Trains a wrapped learner on a given training set, w.r.t. some hyperparameters, case weights and costs. You have to implement this method if you want to add another learner to this package.

Arguments

.learner
[\code\linkS4classlearner]
Wrapped learner from this package.

.targetvar
[character]
Name of the target variable.

.data
[data.frame]
Complete training set.

weights
[numeric]
Case weights, default is 1, which means every case is assigned equal weight. If your learner does not support this, simply ignore this argument.

costs
[matrix]
Misclassification costs, which should be used during training. If your learner does not support this, simply ignore this argument.

...
[any]
Additional parameters, which need to be passed to the underlying train function.

Value

train.learner-methods: Model of the underlying learner.