train.learner,kernlab.svm.classif,formula,data.frame,numeric,list-method(mlr)	R Documentation

Overwritten to allow direct passing of kernel hyperparameters.

Description

Overwritten, to allow direct passing of kernel hyperparameters. Besides that, simply delegates to super method.

Arguments

wrapped.learner 	 Object of class wrapped.learner.

formula 	 A symbolic description of the model to be fitted.

data 	 Dataframe which includes all the data for the task.

weights 	 An optional vector of weights to be used in the fitting process. Default is a weight of 1 for every case.

parset 	 Named list which contains the hyperparameters of the learner. Default is an empty list, which means no hyperparameters are specifically set and defaults of the underlying learner are used.

[Package mlr version 0.3.180 Index]