trainLearner	R Documentation

trainLearner,BaseWrapper,LearnTask,integer-method

Description

Mainly for internal use. Trains a wrapped learner on a given training set, w.r.t. some hyperparameters and case weights. You have to implement this method if you want to add another learner to this package.

Arguments

.learner
[Learner]
Wrapped learner from this package.

.task
[LearnTask]
Task to train learner on.

.subset
[integer]
Subset of cases, index task with this.

...
[any]
Additional parameters, which need to be passed to the underlying train function.

Value

trainLearner-methods: Model of the underlying learner.

See Also

getData