tune	R Documentation

Hyperparameter tuning

Description

Optimizes the hyperparameters of a learner for a classification or regression problem. Allows for different optimization methods, commonly grid search is used but other search techniques are available as well. The specific details of the search algorithm are set by passing a control object.

Usage

tune(learner, task, resampling, control, measures, aggr, model=FALSE,
    path=FALSE)

Arguments

learner
[learner or string]
Learning algorithm. See learners.

task
[learn.task]
Learning task.

resampling
[resample.instance] or [resample.desc]
Resampling strategy to evaluate points in hyperparameter space. At least for grid search, if you pass a description, it is instantiated at one, so all points are evaluated on the same training/test sets.

control
[tune.control]
Control object for search method. Also selects the optimization algorithm for tuning.

measures
[see measures]
Performance measures.

aggr
[see aggregations]
Aggregation functions.

model
[boolean]
Should a final model be fitted on the complete data with the best found hyperparameters? Default is FALSE.

path
[boolean]
Should optimization path be saved? Default is FALSE.

Details

The first measure, aggregated by the first aggregation function is optimized, to find a set of optimal hyperparameters.

Value

opt.result.

See Also

grid.control, optim.control, cmaes.control