values	R Documentation

Get discrete values.

Description

Get discrete values.

Arguments

x
[Parameter | ParameterSet]
A single discrete parameter or a set of parameters.

only.names
[logical(1)]
If TRUE forces the return only value names as strings, otherwise the real values are returned. Default is FALSE.

Value

values-methods: For a single parameter: Its list of values or a character vector if only.names is TRUE. For a Parameterset: A named list of lists of values or a list of character vectors (names of values).