wrapped.learner-class(mlr)	R Documentation

Wrapped...

Description

Wrapped.learner —text!—-

Slots

learner.name:

(character) Name of the learning method

learner.pack:

(character) R package where learner is defined

train.fct:

(function) Function used in above package to train a regular model in the package (see train)

train.fct.pars:

(list) Named list of parameters which are fixed in the above train.fct. NB: These are _NOT_ hyperparamters of the classifier but rather parameters which are fixed for the whole experiment and should be fixed at the beginning for convenience (Example would be the tolerance parameter in lda.)

predict.fct:

(function) Function used in above package to predict new data with a trained model (see predict)

predict.newdata.arg:

(character) Name of argument for the new data frame in the underlying predict method.

predict.fct.pars:

(list) Named list of parameters which are fixed in the above predict.fct. See train.fct.pars (example would be the method parameter in predict.lda.)

learner.props:

(learner.props) Properties of the learner

Methods

train.learner

initialize

signature( = "wrapped.learner"): Constructor.

as.character

signature( = "wrapped.learner"): Conversion to string.

print

signature( = "wrapped.learner"): Prints the object by calling as.character.

show

signature( = "wrapped.learner"): Shows the object by calling as.character.

set.train.par

signature( = "wrapped.learner"): Set a parameter for the underlying train function of a wrapped learner. This is not meant for hyperparamters, pass these through the usual parset argument, but rather to fix (somewhat techical) arguments which stay the same for the whole experiment. You should not have to use this too often.

set.predict.par

signature( = "wrapped.learner"): Set a parameter for the underlying predict function of a wrapped learner. Used to fix (somewhat techical) arguments which stay the same for the whole experiment. Y You should not have to use this too often.

[Package mlr version 0.3.115 Index]