wrapped.learner.classif-class(mlr)	R Documentation

wrapped.learner.classif

Description

Wraps an already implemented classification method from R to make it accesible to mlr.

Slots

train.par.for.classes:

(list) List of parameters to train a model for class predictions. Will be added to train.fct.pars in this case.

train.par.for.probs:

(list) List of parameters to train a model for probability predictions. Will be added to train.fct.pars in this case.

predict.par.for.classes:

(list) List of parameters to predict a model for class predictions. Will be added to predict.fct.pars in this case.

predict.par.for.probs:

(list) List of parameters to predict a model for probability predictions. Will be added to predict.fct.pars in this case.

trafo.for.classes:

(function) Function to transform the raw predictions to classes

trafo.for.probs:

(function) Function to transform the raw predictions to probabilities

dummy.classes:

(logical) Does the predict function need a class column in the newdata dataframe for prediction? If TRUE but no class column is avaible in the data a null column is internally generated (default is FALSE).

Extends

wrapped.learner

Methods

initialize

signature( = "wrapped.learner.classif"): Constructor.

[Package mlr version 0.3.174 Index]