wrapped.model-class	R Documentation

Induced model of learner.

Description

Result from train. It internally stores the underlying fitted model, the IDs of the subset used for training, variables used for training and information about second-level optimization like tuned hyperparameters or selected variables.

Details

wrapped.model-class: Getter.
Note that all getters of task.desc and data.desc can also be used.

learner [learner]

Learner that was used to fit the model.

learner model [any]

Underlying model from used R package.

subset [integer]

Subset used for training.

fail [NULL | string]

Generally NULL but if the training failed, the error message of the underlying train function.

Extends

object

Methods

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

resample.update

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

predict

signature(object = "wrapped.model"): Predict the target variable of new data using a fitted model. If the type is set to "prob" or "decision" probabilities or decision values will be stored in the resulting object. The resulting class labels are the classes with the maximum values or thresholding can also be used.

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

pred.learner

to.string

[

signature( = "wrapped.model"): Getter.