learner.props-class(mlr)	R Documentation

Description object for the features of a learning algorithm.

Description

Description object for the features of a learning algorithm.

Slots

supports.missing:

(logical) [logical] Can missing values be dealt with?

supports.numerics:

(logical) [logical] Are numeric variables ok?

supports.factors:

(logical) [logical] Are factor variables ok?

supports.characters:

(logical) [logical] Are character variables ok?

supports.weights:

(logical) [logical] Can case weights be dealt with?

Methods

[Package mlr version 0.3.180 Index]