resample.desc-class(mlr)	R Documentation

resample.desc

Description

Base class for description of resampling algorithms. A description of a resampling algorithm contains all necessary information to provide a resampling.instance, when given the size of the data set.

Slots

instance.class:

(character) S4 class name of the corresponding resample.instance

name:

(character) Name of this resampling algorithm

iters:

(numeric) Number of iterations

Methods

benchmark

signature(learn.task = "learn.task", outer.resampling = "resample.instance", inner.resampling = "resample.desc", ranges = "list", measure = "list", all.tune.results = "logical"): benchmark conducts a benchmark experiment for a single classifier on a single data set. This consists of an inner stage and outer stage. At the outer stage a tuning set and a test set are repeatedly formed from the data through resampling (usually cross-validation or bootstrapping). The respective hyperparameters of the classifier are tuned on the tuning set again through an inner resampling process, the classifier is trained on the complete tuning set with the best found hyperparameters and the performance is measured on the test set.

[

signature( = "resample.desc"): Getter.

make.resample.instance

signature(desc = "resample.desc", size = "numeric"): Mainly for internal use. Construct a resample.instance from a resample.desc. Convenience method, so one doesn't have to call the specific constructors of classes inheriting from resample.desc.

[Package mlr version 0.3.89 Index]