resample.fit	R Documentation

Fit models according to a resampling strategy.

Description

Given a resampling strategy, which defines sets of training and test indices, resample.fit fits the selected learner using the training sets and performs predictions for the test sets. For construction of the resampling strategies use the factory methods make.res.desc and make.res.instance.

Usage

resample.fit(learner, task, resampling, par.vals, vars, extract)

Arguments

learner
[learner or character]
Learning algorithm.

task
[learn.task]
Learning task.

resampling
[resample.desc or resample.instance]
Resampling strategy.

par.vals
[list]
Named list of hyperparameter values. Will overwrite the ones specified in the learner object. Default is empty list.

vars
[character]
Vector of variable names to use in training the model. Default is to use all variables.

extract
[function]
Function used to extract information from fitted models, e.g. can be used to save the complete list of fitted models. Default is to extract nothing.

Details

resample.fit-methods: Optionally information from the fitted models, e.g. the complete model, can be extracted and returned.

Value

resample.fit-methods: resample.prediction.