wrapped.learner-class(mlr)	R Documentation

wrapped.learner

Description

Wraps an already implemented learning method from R to make it accesible to mlr.

Details

Also inlcudes a properties object to describe the features of the learner.

Slots

learner.name:

(character) Descriptive name of the learning method

learner.pack:

(character) R package where learner is implemented

train.fct:

(function) Function used in above package to train a regular model in the package

train.fct.pars:

(list) Named list of parameters which are fixed in the above train.fct and used at every internal call.

predict.fct:

(function) Function used in above package to predict new data with a trained model

predict.newdata.arg:

(character) Name of argument for the new data frame in the underlying predict method.

predict.fct.pars:

(list) Named list of parameters which are fixed in the above predict.fct and used at every internal call.

learner.props:

(learner.props) Properties of the learner

Methods

train.learner

signature(wrapped.learner = "wrapped.learner", formula = "formula", data = "data.frame", weights = "numeric", parset = "list"): Mainly for internal use. Trains a wrapped learner on a giving training set, possibly w.r.t. some hyperparamters and case weights.

initialize

signature( = "wrapped.learner"): Constructor.

set.train.par

signature( = "wrapped.learner"): Set a parameter for the underlying train function of a [wrapped.learner]. This is not meant for hyperparameters, pass these through the usual parset argument, but rather to fix (somewhat technical) arguments which stay the same for the whole experiment. You should not have to use this too often.

set.predict.par

signature( = "wrapped.learner"): Set a parameter for the underlying predict function of a wrapped learner. Used to fix (somewhat techical) arguments which stay the same for the whole experiment. You should not have to use this too often.

[Package mlr version 0.3.174 Index]