wrapped.model-class(mlr)	R Documentation

wrapped.model

Description

A classifier model is generated by train. It consists of the learn.task, the classification model, the used hyperparameters and a vector specifying the cases of the training sample.

Slots

task.class:

(character) [learn.task] Specifies learning task

learner.class:

(character) [character] Specifies learner class

learner.name:

(character) [character] Name of the learning method

learner.model:

(ANY) [external.model] External model from existing R packages like lda, rpart, etc

subset:

(numeric) [integer] An index vector specifying the cases of the training sample that were used for the model.

vars:

(character) [charactter] Vector specifying the variables which were used to build the model.

parset:

(list) [list] Contains the hyperparameters of the train function. If empty no parameters were used.

Methods

as.character

signature( = "wrapped.model"): Conversion to string.

print

signature( = "wrapped.model"): Prints the object by calling as.character.

show

signature( = "wrapped.model"): Shows the object by calling as.character.

[

Examples

see train

[Package mlr version 0.3.115 Index]