wrapped.model-class(mlr)	R Documentation

wrapped.model

Description

A classifier model is generated by train. It consists of the learn.task, the classification model, the used hyperparameters and a vector specifying the cases of the training sample.

Slots

task.class:

(character) String

learner.class:

(character) Specifies learner class

learner.name:

(character) Name of the learning method

learner.model:

(ANY) External model from existing R packages like lda, rpart, etc

subset:

(numeric) An index vector specifying the cases of the training sample that were used for the model.

vars:

(character) Vector specifying the variables which were used to build the model.

parset:

(list) Contains the hyperparameters of the train function. If empty no parameters were used.

Methods

[

signature( = "wrapped.model"): Getter.

Examples

see train

[Package mlr version 0.3.180 Index]